requirements.txt
Django==4.1.8
dj_database_url==1.3.0
cryptography==40.0.1
pyasn1-modules==0.2.8
django-environ==0.10.0
#psycopg2==2.9.6
gunicorn==20.1.0
cryptography
PyJWT
django-allauth